refactor: User 관련 패키지 구조 변경 및 경로 수정
1. UserController의 경로를 presentation에서 adapter.in.web으로 이동한다.
2. UserCreateRequest 및 UserInfoResponse의 패키지를 presentation.dtos에서 application.commandquery로 변경한다.
3. UsersService 및 관련 엔티티, 모델을 반영하여 경로를 수정한다.

Resolves #22
