+++ src/main/java/com/posicube/assignment/users/adapter/in/web/UserController.java
@@ -1,10 +1,11 @@
-package com.posicube.assignment.users.presentation;
+package com.posicube.assignment.users.adapter.in.web;
 
 import com.posicube.assignment.common.baseResponse.BaseResponse;
-import com.posicube.assignment.users.application.UserService;
-import com.posicube.assignment.users.domain.entity.Users;
-import com.posicube.assignment.users.presentation.dtos.UserCreateRequest;
-import com.posicube.assignment.users.presentation.dtos.UserInfoResponse;
+import com.posicube.assignment.users.application.service.UsersService;
+import com.posicube.assignment.users.adapter.out.persistence.UsersJpaEntity;
+import com.posicube.assignment.users.application.commandquery.UserCreateRequest;
+import com.posicube.assignment.users.application.commandquery.UserInfoResponse;
+import com.posicube.assignment.users.domain.model.Users;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -18,11 +19,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/users")
 @RequiredArgsConstructor
 public class UserController {
-    private final UserService userService;
+    private final UsersService usersService;
 
     @PostMapping("")
     public BaseResponse<UserInfoResponse> create(@Valid @RequestBody UserCreateRequest request) {
-        Users user = userService.createUser(request);
+        Users user = usersService.createUser(request);
         return new BaseResponse<>(UserInfoResponse.from(user));
     }
 }
